+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 2026-06-24 (9.11.0)
+
+* Added soft and hard delete functionalities to delete tables after grace period.
+
 # 2026-06-22 (9.10.2)
 
 * Bugfix: output the leaf nodes of the ErrorTree rather than the root in case of anyOf/oneOf